backend/requirements.txt
Django==3.2.3
django-filter==22.1
djangorestframework==3.12.4
drf-base64==2.0
djoser==2.1.0
webcolors==1.11.1
Markdown==3.3.4
Pillow==9.2.0
python-dotenv==0.20.0
gunicorn==20.1.0
flake8==5.0.4
reportlab==3.6.11
PyYAML==6.0
psycopg2-binary==2.9.9
python-decouple-3.8